Fix the build against current glibc and GCC
Four unrelated defects, all of which stop the tree compiling on a modern
toolchain. None are caused by a change in this branch.

utils/config.h gated HAVE_STRCASESTR and HAVE_STRCHRNUL on _GNU_SOURCE
alone. glibc guards both declarations with __USE_MISC, which _DEFAULT_SOURCE
and _BSD_SOURCE also enable, so a build that defined either of those but not
_GNU_SOURCE redeclared the functions and tripped -Werror=redundant-decls.
Test the same set of macros glibc does. This is in the shared header rather
than per-frontend because every target routes through it.

frontends/monkey/browser.c had a genuine buffer overflow. The argc < 5 guard
in monkey_window_handle_exec reported "WINDOW EXEC ARGS BAD" but did not
return, so strcpy(cmd, argv[4]) read past the end of argv and wrote into a
zero-byte calloc. monkey_window_handle_click had the same missing return.
Those were the only two of the ten ARGS BAD sites affected.

Thirteen calloc calls passed their arguments transposed as
calloc(sizeof(x), n). Harmless in practice but diagnosed by
-Werror=calloc-transposed-args on GCC 14.

Two curl_easy_setopt calls passed int where the varargs contract requires
long, which the curl typecheck macros reject.

+++ netsurf/content/fetchers/curl.c
@@ -1206,7 +1206,7 @@ static CURLcode fetch_curl_set_options(struct curl_fetch_info *f)
 	}
 
 	if ((auth = urldb_get_auth_details(f->url, NULL)) != NULL) {
-		SETOPT(CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
+		SETOPT(CURLOPT_HTTPAUTH, (long) CURLAUTH_BASIC);
 		SETOPT(CURLOPT_USERPWD, auth);
 	} else {
 		SETOPT(CURLOPT_USERPWD, NULL);
@@ -1243,7 +1243,7 @@ static CURLcode fetch_curl_set_options(struct curl_fetch_info *f)
 	}
 
 	/* Force-enable SSL session ID caching, as some distros are odd. */
-	SETOPT(CURLOPT_SSL_SESSIONID_CACHE, 1);
+	SETOPT(CURLOPT_SSL_SESSIONID_CACHE, 1L);
 
 	if (urldb_get_cert_permissions(f->url)) {
 		/* Disable certificate verification */
